Enable customization of plots
Introduces objects to encapsulate all the backend-specific objects that a user
might need to access in order to customize plots.

This commit focuses on providing access to the Figure and Axes objects from
matplotlib, as these have been specifically requested. Access to additional
objects is expected to be added in future commits.

Signed-off-by: John Pennycook <john.pennycook@intel.com>