Drop duplicate (application, platform) results
All of our existing tests and case studies assume that the dataset contains
only a single performance result for a given (application, platform) pair.
In the event that multiple results exist (e.g., because of combining results
during projection) various aspects of the metrics and plotting scripts start to
misbehave.

This commit fixes this issue by selecting the (application, platform) pairs
with the highest efficiencies prior to computing performance portability or
plotting cascades.

Signed-off-by: John Pennycook <john.pennycook@intel.com>