Update coverage schema to 0.2.0
The previous schema did not allow for coverage files to contain path
information. While a file hash is sufficient to identify uniqueness when
computing divergence, a path is more useful to humans and downstream tools.

The new schema allows for coverage files to contain both a file hash
and (optional) path information.

Signed-off-by: John Pennycook <john.pennycook@intel.com>