Simplify divergence usage by making cov optional
Previously, invoking the divergence function (and any functions that invoke
the divergence function, like snapshot) required the P3 and coverage
information to be stored in separate tables and connected by a key.

This setup is convenient for data drawn from a database that uses this
representation already, but is inconvenient for users reading in data
that directly associates a coverage string with each performance result.

This commit makes the cov DataFrame optional: if it exists, then behavior
is the same as before; if it does not exist, then the library looks for
coverage information stored alongside the P3 data.

Signed-off-by: John Pennycook <john.pennycook@intel.com>