Avoid false positives in _cast_to_numeric
Previously, _cast_to_numeric would raise a TypeError if a column was
missing. We now check that a column exists before trying to cast it.

This does result in some redundant checking, as some functions call both
_require_columns and _cast_to_numeric. We may want to consider
refactoring to combine all of this functionality into a single DataFrame
validation step.

Signed-off-by: John Pennycook <john.pennycook@intel.com>